+++ crates/burn-cube/src/language/element/mod.rs
@@ -1,19 +1,18 @@
 mod array;
 mod base;
 mod bool;
-mod conversion;
+mod cast;
+mod cube_elem;
 mod float;
 mod int;
 mod numeric;
-mod primitive;
-mod static_type;
 mod uint;
 
 pub use array::*;
 pub use base::*;
-pub use conversion::*;
+pub use cast::*;
+pub use cube_elem::*;
 pub use float::*;
 pub use int::*;
 pub use numeric::*;
-pub use primitive::*;
 pub use uint::*;
