--- crates/burn-cube/src/language/element/static_type.rs
@@ -1,18 +1,0 @@
-use crate::language::CubeType;
-
-/// Types that exist within a cube function
-/// but should not be turned into a JIT variable
-/// For instance: a bool that determines at compile time
-/// if we should unroll a for loop or not
-
-impl CubeType for u32 {
-    type ExpandType = u32;
-}
-
-impl CubeType for f32 {
-    type ExpandType = f32;
-}
-
-impl CubeType for i32 {
-    type ExpandType = i32;
-}
